+++ src/OpenTelemetry.Exporter.Console/ConsoleExporterHelperExtensions.cs
@@ -15,9 +15,7 @@
 // </copyright>
 
 using System;
-using System.Diagnostics;
 using OpenTelemetry.Exporter;
-using OpenTelemetry.Exporter.Console;
 
 namespace OpenTelemetry.Trace
 {
@@ -39,7 +37,7 @@ namespace OpenTelemetry.Trace
 
             var options = new ConsoleExporterOptions();
             configure?.Invoke(options);
-            return builder.AddProcessor(new SimpleExportProcessor<Activity>(new ConsoleActivityExporter(options)));
+            return builder.AddProcessor(new SimpleActivityExportProcessor(new ConsoleActivityExporter(options)));
         }
     }
 }