+++ src/OpenTelemetry.Exporter.Console/ConsoleExporterLoggingExtensions.cs
@@ -17,7 +17,6 @@
 #if NET461 || NETSTANDARD2_0
 using System;
 using OpenTelemetry.Exporter;
-using OpenTelemetry.Exporter.Console;
 
 namespace OpenTelemetry.Logs
 {
@@ -38,7 +37,7 @@ namespace OpenTelemetry.Logs
 
             var options = new ConsoleExporterOptions();
             configure?.Invoke(options);
-            return loggerOptions.AddProcessor(new SimpleExportProcessor<LogRecord>(new ConsoleLogRecordExporter(options)));
+            return loggerOptions.AddProcessor(new SimpleLogRecordExportProcessor(new ConsoleLogRecordExporter(options)));
         }
     }
 }